+++ header.php
@@ -18,7 +18,7 @@
         <link rel="stylesheet" type="text/css" href="<?php bloginfo('template_directory');?>/style.css">
         <link rel="shortcut icon" href="<?php bloginfo('template_directory');?>/images/icons/favicon.ico" type="image/x-icon">
         <!--[if lt IE 9]>
-        <script src="<?php echo get_template_directory_uri(); ?>/js/html5.js"></script>
+            <script src="http://html5shiv.googlecode.com/svn/trunk/html5.js"></script>
         <![endif]-->
 
         <?php