+++ CHANGELOG.md
@@ -4,4 +4,12 @@ CHANGELOG
 21/03/2014
 ----------
 
-Initial release
+* Initial release
+
+
+22/03/2014
+----------
+
+* Using google hosted html5shyv for better performance
+* jQuery UI and jQuery mobile support (disabled by default)
+* jQuery updated to 1.11.0